rename watch command
Переименовал команду watch, чтобы он регистрировалась первая
